requirements.txt
# Flask 기본 패키지
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Session==0.5.0

# 데이터베이스
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Redis 캐시
redis==5.0.1

# 배치 스케줄러 (추가)
APScheduler==3.10.4
pytz==2023.3

# HTTP 요청 (ERPia API)
requests==2.31.0

# 날짜/시간 처리
python-dateutil==2.8.2

# 웹 서버
gunicorn==21.2.0

# 개발 도구
python-dotenv==1.0.0 